+++ src/controllers/Game.js
@@ -578,5 +578,5 @@ exports.isCurrentPlaySecondVoteAfterTie = async game => {
     const previousPlay = await GameHistory.getPreviousPlay(game._id);
     const currentPlay = game?.waiting.length ? game.waiting[0] : undefined;
     return currentPlay?.to === "vote" && previousPlay?.play.action === "vote" &&
-        previousPlay.turn === game.turn && previousPlay.play.targets.length > 1;
+        previousPlay.turn === game.turn && previousPlay.play.votesResult === "need-settlement";
 };